Move video into videoDevice file/capability
